requirements_core.txt
streamlit==1.29.0
pandas==2.0.3
psycopg2-binary==2.9.7